% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{analyse_communities}
\alias{analyse_communities}
\title{Analyse communities function
Prints summary information about the communities}
\usage{
analyse_communities(g, communities, verbose = FALSE)
}
\description{
Analyse communities function
Prints summary information about the communities
}
