% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus.R
\name{consensus_communities}
\alias{consensus_communities}
\title{calculates consensus based on co-occurrence matrix}
\usage{
consensus_communities(nco, gamma_lim)
}
\description{
calculates consensus based on co-occurrence matrix
}
