% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{empirical_mu}
\alias{empirical_mu}
\title{Empirical Mu function
Calculates the value of mixing parameter "mu"}
\usage{
empirical_mu(g)
}
\description{
Empirical Mu function
Calculates the value of mixing parameter "mu"
}
