% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_comms.R
\name{find_communities}
\alias{find_communities}
\title{Basic function to find communities}
\usage{
find_communities(g, method, r = c(1), s = c(10), IMtrials = 1, verbose = FALSE)
}
\description{
Basic function to find communities
}
