% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus.R
\name{find_communities_N_times}
\alias{find_communities_N_times}
\title{community detection, repeated (THIS FUNCTION IS DEPRECATED)}
\usage{
find_communities_N_times(
  g,
  n_trials,
  methods,
  shuffle = FALSE,
  resolution = c(1),
  filename_summary = "",
  filename_membership = ""
)
}
\description{
community detection, repeated (THIS FUNCTION IS DEPRECATED)
}
