% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus.R
\name{find_communities_repeated}
\alias{find_communities_repeated}
\title{community detection, repeated t times}
\usage{
find_communities_repeated(
  g,
  n_trials,
  method = method,
  shuffle = TRUE,
  resolution = c(1),
  verbose = FALSE
)
}
\description{
community detection, repeated t times
}
