% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{make_community_network}
\alias{make_community_network}
\title{This function builds a network (Gc) whose nodes represent communities within the original network (G).
Gc can be utilized for in-depth analysis and visualization of the relationships among these communities.}
\usage{
make_community_network(g)
}
\arguments{
\item{g:}{The network to be analyzed. It should be an iGraph object with specific attributes:
\itemize{
\item V(g)$community: An integer value representing the community assignment for each node.
\item E(g)$w: A numeric vector containing edge weights. If the network is unweighted, set E(g)$w to 1.0.
}}
}
\value{
A network object (Gc) with the following attributes:
\itemize{
\item $membership: Stores the community labels.
\item E(Gc)$w: The sum of corresponding edge weights in the original network (G).
\item V(Gc)$size: the number of nodes of the original network G that are represented in a node of Gc.
}
}
\description{
This function builds a network (Gc) whose nodes represent communities within the original network (G).
Gc can be utilized for in-depth analysis and visualization of the relationships among these communities.
}
