% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus.R
\name{normalized_co_occurrence}
\alias{normalized_co_occurrence}
\title{calculates normalized co occurence matrix
X is a matrix of x_ij counts of co-occurrence
X_normalized is a matrix of gamma_ij normalized coefficients
expressing the probability that i and j are in the same community
X_normalized_ij == 1.0 means that i and j have ALWAYS been classified in the same community
X_normalized_ij == 0.0 means that i and j have NEVER been classified in the same community}
\usage{
normalized_co_occurrence(M)
}
\description{
calculates normalized co occurence matrix
X is a matrix of x_ij counts of co-occurrence
X_normalized is a matrix of gamma_ij normalized coefficients
expressing the probability that i and j are in the same community
X_normalized_ij == 1.0 means that i and j have ALWAYS been classified in the same community
X_normalized_ij == 0.0 means that i and j have NEVER been classified in the same community
}
