% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus.R
\name{find_communities}
\alias{find_communities}
\title{community detection,
retunrs a community object
including the algorithm
applies selected method
applies resolution for LV and LD
undirected(g) for LV and LD}
\usage{
find_communities(g, method, r = c(1), verbose = FALSE)
}
\description{
community detection,
retunrs a community object
including the algorithm
applies selected method
applies resolution for LV and LD
undirected(g) for LV and LD
}
