% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{make_community_network}
\alias{make_community_network}
\title{Make Community Network function
returns a network GC whose nodes are the communities of the original network G
may be useful to analyse and visualise the relationships among communities}
\usage{
make_community_network(g)
}
\description{
Make Community Network function
returns a network GC whose nodes are the communities of the original network G
may be useful to analyse and visualise the relationships among communities
}
